[librpbase] byteswap.h: Use `size_t` instead of `unsigned int` for sizes.
Fixes a few conversion warnings in 64-bit builds.

Note that this slightly increases the code size in 64-bit builds
for the C++ implementation and the SSE2 version.

gcc-9.2.0, 64-bit Linux build:

Using `unsigned int`:
   text    data     bss     dec     hex filename
    389       0       0     389     185 byteswap.c.o
    289       0       0     289     121 byteswap_ifunc.c.o
    449       0       0     449     1c1 byteswap_sse2.c.o
    417       0       0     417     1a1 byteswap_ssse3.c.o

Using `size_t`:
   text    data     bss     dec     hex filename
    413       0       0     413     19d byteswap.c.o
    289       0       0     289     121 byteswap_ifunc.c.o
    457       0       0     457     1c9 byteswap_sse2.c.o
    409       0       0     409     199 byteswap_ssse3.c.o

+++ src/librpbase/byteswap.c
@@ -28,7 +28,7 @@ static inline uint32_t swap_two_16_in_32(uint32_t dword)
  * @param ptr Pointer to array to swap. (MUST be 16-bit aligned!)
  * @param n Number of bytes to swap. (Must be divisible by 2; an extra odd byte will be ignored.)
  */
-void __byte_swap_16_array_c(uint16_t *ptr, unsigned int n)
+void __byte_swap_16_array_c(uint16_t *ptr, size_t n)
 {
 	uint32_t *dwptr;
 
@@ -68,7 +68,7 @@ void __byte_swap_16_array_c(uint16_t *ptr, unsigned int n)
  * @param ptr Pointer to array to swap. (MUST be 32-bit aligned!)
  * @param n Number of bytes to swap. (Must be divisible by 4; extra bytes will be ignored.)
  */
-void __byte_swap_32_array_c(uint32_t *ptr, unsigned int n)
+void __byte_swap_32_array_c(uint32_t *ptr, size_t n)
 {
 	// Verify the block is 32-bit aligned
 	// and is a multiple of 4 bytes.
